requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.18.0
amqp==5.2.0
anyio==3.6.2
apischema==0.18.0
APScheduler==3.10.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
autopep8==2.0.1
Babel==2.8.0
backcall==0.2.0
backends==1.4.32
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
bleach==5.0.1
blinker==1.4
boto3==1.24.27
botocore==1.27.96
Brlapi==0.8.3
bs4==0.0.1
CacheControl==0.12.11
celery==5.2.7
celery-pool-asyncio==0.2.0
celery-redbeat==2.1.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
crashtest==0.4.1
cron-descriptor==1.2.35
cryptography==41.0.3
cupshelpers==1.0
DateTime==5.1
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==0.5.0
Django==4.1.7
django-admin-honeypot-2==1.1.0
django-allauth==0.51.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-ckeditor==6.7.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-honeypot==1.1.0
django-ipware==6.0.3
django-js-asset==2.1.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django-storages==1.12.3
django-timezone-field==5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
djangorestframework-xml==2.0.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.5
dulwich==0.20.50
duplicity==0.8.21
ecdsa==0.18.0
email-validator==1.3.1
entrypoints==0.4
environ==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
Faker==21.0.0
fastapi==0.95.0
fasteners==0.14.1
fastjsonschema==2.16.2
fdm==0.4.1
filelock==3.9.0
Flask==1.1.4
Flask-Cors==3.0.10
frozenlist==1.2.0
future==0.18.2
gevent==22.10.2
google-cloud==0.34.0
gpg==1.16.0
greenlet==2.0.2
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httplib2==0.20.2
httpx==0.24.0
idna==3.4
impass==0.12
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.11
ipfshttpclient==0.7.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==1.1.0
itypes==1.2.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keyring==23.13.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.6.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.4
multiaddr==0.0.9
multidict==5.1.0
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
olefile==0.46
openai==0.28.0
openapi-codec==1.3.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.5.4
pluggy==1.0.0
plum-dispatch==1.7.4
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.38
protobuf==3.12.4
proton-core==0.1.16
proton-keyring-linux==0.0.1
proton-keyring-linux-secretservice==0.0.1
proton-vpn-api-core==0.20.2
proton-vpn-connection==0.11.0
proton-vpn-gtk-app==4.1.3
proton-vpn-killswitch==0.2.0
proton-vpn-killswitch-network-manager==0.2.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.3.0
proton-vpn-network-manager-openvpn==0.0.4
proton-vpn-session==0.6.4
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pyftpdlib==1.5.7
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.2
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.3.0
python-apt==2.4.0+ubuntu2
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-gnupg==0.4.8
python-ipware==2.0.1
python-jose==3.3.0
python-multipart==0.0.6
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.7.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rapidfuzz==2.13.7
redis==4.5.1
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
schedule==1.1.0
scipy==1.10.0
SecretStorage==3.3.1
selenium==4.15.2
Send2Trash==1.8.0
shellingham==1.5.0.post1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
solenoid==0.3.1
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
sqlparse==0.4.3
stack-data==0.6.1
starlette==0.26.1
sympy==1.11.1
systemd-python==234
tabulate==0.8.9
tenacity==8.2.3
terminado==0.17.0
texttable==1.6.4
tinycss2==1.2.1
tk==0.1.0
tkintertable==1.3.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.64.0
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.1.20
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
uvicorn==0.21.1
validate-docbr==1.10.0
varint==1.0.2
vine==5.0.0
virtualenv==20.18.0
wadllib==1.3.6
wcwidth==0.2.6
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==1.0.1
whitenoise==5.3.0
widgetsnbextension==4.0.3
win32-setctime==1.1.0
wsproto==1.2.0
xdg==5
xdo==0.5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
zope.event==4.6
zope.interface==6.0
